+++ app/src/main/java/ee/ria/DigiDoc/android/Application.java
@@ -70,6 +70,7 @@ import ee.ria.DigiDoc.android.utils.Formatter;
 import ee.ria.DigiDoc.android.utils.LocaleService;
 import ee.ria.DigiDoc.android.utils.navigator.Navigator;
 import ee.ria.DigiDoc.android.utils.navigator.conductor.ConductorNavigator;
+import ee.ria.DigiDoc.configuration.util.useragent.UserAgentUtil;
 import ee.ria.DigiDoc.configuration.ConfigurationConstants;
 import ee.ria.DigiDoc.configuration.ConfigurationManager;
 import ee.ria.DigiDoc.configuration.ConfigurationManagerService;
@@ -163,7 +164,8 @@ public class Application extends android.app.Application {
     // Container configuration
 
     private void setupSignLib() {
-        SignLib.init(this, getString(R.string.main_settings_tsa_url_key), getConfigurationProvider());
+        UserAgentUtil userAgent = new UserAgentUtil(getApplicationContext());
+        SignLib.init(this, getString(R.string.main_settings_tsa_url_key), getConfigurationProvider(), userAgent.getUserAgent());
     }
 
     private void setupRxJava() {
