+++ configuration-lib/src/main/java/ee/ria/DigiDoc/configuration/loader/CentralConfigurationClient.java
@@ -1,5 +1,7 @@
 package ee.ria.DigiDoc.configuration.loader;
 
+import android.content.Context;
+
 import java.io.IOException;
 import java.security.KeyStore;
 import java.security.SecureRandom;
@@ -10,23 +12,26 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSocketFactory;
 import javax.net.ssl.TrustManagerFactory;
 
+import ee.ria.DigiDoc.configuration.util.useragent.UserAgentUtil;
 import okhttp3.Call;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.ResponseBody;
 
-class CentralConfigurationClient {
+public class CentralConfigurationClient {
 
     private static final int DEFAULT_TIMEOUT = 5;
     private final OkHttpClient httpClient;
     private final String centralConfigurationServiceUrl;
     private final X509Certificate explicitSSLCert;
+    private final Context context;
 
-    CentralConfigurationClient(String centralConfigurationServiceUrl, X509Certificate explicitSSLCert) {
+    CentralConfigurationClient(String centralConfigurationServiceUrl, X509Certificate explicitSSLCert, Context context) {
         this.centralConfigurationServiceUrl = centralConfigurationServiceUrl;
         this.explicitSSLCert = explicitSSLCert;
         httpClient = constructHttpClient();
+        this.context = context;
     }
 
     String getConfiguration() {
@@ -42,9 +47,13 @@ class CentralConfigurationClient {
     }
 
     private String requestData(String url) {
+
+        UserAgentUtil userAgentUtil = new UserAgentUtil(context);
+
         Request request = new Request.Builder()
                 .url(url)
                 .addHeader("Content-Type", "application/json")
+                .addHeader("User-Agent", userAgentUtil.getUserAgent())
                 .build();
 
         Call call = httpClient.newCall(request);
